Create Check-Exchange-NTFS-ClusterSize-1.0.ps1
Purpose of this script to review the NTFS allocation size on all disks which are attached to Exchange servers.  WMI is used to query for the the volume information.
This allows it to query remote servers.  There is logic to loop through multiple Exchange servers, and all the disks attached to each serve